rename gast_to_code as gast_to_code_main
